Implement echo -quoting json
Format each argument as a json string and join on newline

Split out quoting.hh from string_utils.hh. Also split
options_strings.hh from string_utils.hh. This seemed
less awkward than keeping the material in quoting.hh.

Added `#include <algorithm>` to strings.hh. Without this compile
failed on FreeBSD 14.1, clang version 18.1.5.
