+++ src/normal.cc
@@ -14,7 +14,9 @@
 #include "file.hh"
 #include "flags.hh"
 #include "option_manager.hh"
+#include "option_strings.hh"
 #include "option_types.hh"
+#include "quoting.hh"
 #include "ranges.hh"
 #include "regex.hh"
 #include "register_manager.hh"
